Now you can delete post without reloading and it is fast too

+++ src/pages/Home.js
@@ -14,8 +14,12 @@ import boxicons from 'boxicons';
 function Home({ isAuth }) {
   const [postLists, setPostList] = useState([]);
   const deletePost = async (id) => {
+    toast.success("Post Deleted Successfully", {
+      autoClose: 1500,
+    });
     await deletePostFromDb(id);
-    window.location.reload();
+    const updatedPostList = postLists.filter(post => post.id !== id);
+    setPostList(updatedPostList);
   };
 
   // const DUMMY_POST = {
@@ -39,7 +43,7 @@ function Home({ isAuth }) {
   const sharingHandler = (s) => {
     // console.log(`https://blogweet.vercel.app${s}`);
     navigator.clipboard.writeText(`https://blogweet.vercel.app${s}`);
-    toast(`Your link has been pasted to your Clipboard. Enjoy!`);
+    toast.success(`Your link has been pasted to your Clipboard. Enjoy!`);
   };
 
   const [showScrollToTop, setShowScrollToTop] = useState(false);
@@ -72,7 +76,9 @@ function Home({ isAuth }) {
                     className="title"
                     onClick={() => {
                       navigate(
-                        `/${post.author.name.replaceAll(" ", "-")}/${post.id}`,
+                        `/user/${post.author.name.replaceAll(" ", "-")}/${
+                          post.id
+                        }`,
                         { state: post }
                       );
                     }}
@@ -89,7 +95,12 @@ function Home({ isAuth }) {
                     post.author.id === auth.currentUser.uid && (
                       <button
                         onClick={() => {
-                          deletePost(post.id);
+                          const confirmed = window.confirm(
+                            "Are you sure you want to delete this post?"
+                          );
+                          if (confirmed) {
+                            deletePost(post.id);
+                          }
                         }}
                       >
                         {" "}
@@ -102,7 +113,9 @@ function Home({ isAuth }) {
                   <button
                     onClick={() =>
                       sharingHandler(
-                        `/${post.author.name.replaceAll(" ", "-")}/${post.id}`
+                        `/user/${post.author.name.replaceAll(" ", "-")}/${
+                          post.id
+                        }`
                       )
                     }
                   >
@@ -139,7 +152,9 @@ function Home({ isAuth }) {
                     }}
                     onClick={() => {
                       navigate(
-                        `/${post.author.name.replaceAll(" ", "-")}/${post.id}`,
+                        `/user/${post.author.name.replaceAll(" ", "-")}/${
+                          post.id
+                        }`,
                         { state: post }
                       );
                     }}
@@ -153,7 +168,7 @@ function Home({ isAuth }) {
                 <div
                   style={{ cursor: "pointer" }}
                   onClick={() => {
-                    navigate(`/${post.author.name.replaceAll(" ", "-")}`, {
+                    navigate(`/user/${post.author.name.replaceAll(" ", "-")}`, {
                       state: post.author,
                     });
                   }}
@@ -161,22 +176,22 @@ function Home({ isAuth }) {
                   👤{post.author.name}
                 </div>
               </h3>
-              <ToastContainer
-                position="top-center"
-                autoClose={5000}
-                hideProgressBar={false}
-                newestOnTop
-                closeOnClick
-                rtl={false}
-                pauseOnFocusLoss
-                draggable
-                pauseOnHover
-                theme="dark"
-              />
             </div>
           );
         })}
       </div>
+      <ToastContainer
+        position="top-center"
+        autoClose={5000}
+        hideProgressBar={false}
+        newestOnTop
+        closeOnClick
+        rtl={false}
+        pauseOnFocusLoss
+        draggable
+        pauseOnHover
+        theme="dark"
+      />
       {/* Scroll-to-top button */}
       {showScrollToTop && (
         <button
